cache-server/src/requirements.txt
fastapi
pydantic
uvicorn
pipenv
couchbase
websockets
prometheus_client
python-dotenv
starlette
starlette_prometheus
redis